Added name to input field (#28)
Added name to input field in hopes that it will give the possibility to store the key in browser memory and thus suggest it the next time you see the popup. Doesn't help if you nuke the whole browser cache, but might help if you clear local storage specifically. And shouldn't hurt.