increase license-box z-index to keep it visible in more cases
In the apps that have elements with non-zero z-index in the top of the page such elements would overlap and hide the license box even if the app developers do not intend to do so (see Bakery for Flow for example). Setting the z-index of the license-box to a somewhat high positive value leaves more space for app developers while keeping the license warning visible.