Added a longer message and new style. (#62)
Earlier the dialog said: "Click to validate your Vaadin Subscription?"

Update it to say: "This application is using components which are part of a Vaadin subscription. 
Click here to get a trial or validate your subscription"

This will give a better description on what this is more about, and that you can get a trial.